backend/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
python-decouple
chromadb
pypdf2